delete file

+++ src/LogFS/Header.h
@@ -12,7 +12,7 @@ struct LogFSHeader {
   uint8_t version;
   uint16_t sectorSize;
   uint16_t pageSize;
-  uint16_t sectorsAmount;
+  uint32_t sectorsAmount;
   uint32_t sectorsMapStartAddress;
   uint32_t sectorsStartAddress;
